Fix database. Move to sync postgresql.
