+++ src/presentation/telegram/start.py
@@ -13,13 +13,12 @@ start_router = Router()
 async def start(
     message: Message
 ) -> None:
-    
-    async for session in get_session():
-        user_service.add_user(
-            session=session,
-            telegram_id=message.from_user.id
-        )
 
-        await message.answer(
-            text="/testpool"
-        )
+    user_service.add_user(
+        session=get_session(),
+        telegram_id=message.from_user.id
+    )
+
+    await message.answer(
+        text="/testpool"
+    )
